requirements.txt
asgiref==3.5.2
bcrypt==4.0.0
cffi==1.15.1
crispy-bootstrap5==0.6
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
pillow-heif==0.7.0
psycopg2-binary==2.9.3
pycparser==2.21
python-decouple==3.6
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0